+++ hypershift-operator/controllers/hostedcluster/hostedcluster_controller.go
@@ -46,9 +46,7 @@ import (
 	"github.com/openshift/hypershift/hypershift-operator/controllers/manifests"
 	"github.com/openshift/hypershift/hypershift-operator/controllers/manifests/clusterapi"
 	"github.com/openshift/hypershift/hypershift-operator/controllers/manifests/controlplaneoperator"
-	"github.com/openshift/hypershift/hypershift-operator/controllers/manifests/egressfirewall"
 	"github.com/openshift/hypershift/hypershift-operator/controllers/manifests/ignitionserver"
-	"github.com/openshift/hypershift/hypershift-operator/controllers/manifests/networkpolicy"
 	"github.com/openshift/hypershift/support/capabilities"
 	"github.com/openshift/hypershift/support/certs"
 	"github.com/openshift/hypershift/support/config"
@@ -3246,109 +3244,6 @@ func (r *HostedClusterReconciler) reconcileMachineApprover(ctx context.Context, 
 	return machineapprover.ReconcileMachineApprover(ctx, r.Client, hcp, machineApproverImage, utilitiesImage, createOrUpdate, r.SetDefaultSecurityContext, config.MutatingOwnerRefFromHCP(hcp, releaseVersion))
 }
 
-func (r *HostedClusterReconciler) reconcileNetworkPolicies(ctx context.Context, createOrUpdate upsert.CreateOrUpdateFN, hcluster *hyperv1.HostedCluster, hcp *hyperv1.HostedControlPlane) error {
-	controlPlaneNamespaceName := manifests.HostedControlPlaneNamespace(hcluster.Namespace, hcluster.Name).Name
-
-	// Reconcile openshift-ingress Network Policy
-	policy := networkpolicy.OpenshiftIngressNetworkPolicy(controlPlaneNamespaceName)
-	if _, err := createOrUpdate(ctx, r.Client, policy, func() error {
-		return reconcileOpenshiftIngressNetworkPolicy(policy)
-	}); err != nil {
-		return fmt.Errorf("failed to reconcile ingress network policy: %w", err)
-	}
-
-	// Reconcile same-namespace Network Policy
-	policy = networkpolicy.SameNamespaceNetworkPolicy(controlPlaneNamespaceName)
-	if _, err := createOrUpdate(ctx, r.Client, policy, func() error {
-		return reconcileSameNamespaceNetworkPolicy(policy)
-	}); err != nil {
-		return fmt.Errorf("failed to reconcile same namespace network policy: %w", err)
-	}
-
-	// Reconcile KAS Network Policy
-	var managementClusterNetwork *configv1.Network
-	if r.ManagementClusterCapabilities.Has(capabilities.CapabilityNetworks) {
-		managementClusterNetwork = &configv1.Network{ObjectMeta: metav1.ObjectMeta{Name: "cluster"}}
-		if err := r.Get(ctx, client.ObjectKeyFromObject(managementClusterNetwork), managementClusterNetwork); err != nil {
-			return fmt.Errorf("failed to get management cluster network config: %w", err)
-		}
-	}
-	policy = networkpolicy.KASNetworkPolicy(controlPlaneNamespaceName)
-	if _, err := createOrUpdate(ctx, r.Client, policy, func() error {
-		return reconcileKASNetworkPolicy(policy, hcluster, r.ManagementClusterCapabilities.Has(capabilities.CapabilityDNS), managementClusterNetwork)
-	}); err != nil {
-		return fmt.Errorf("failed to reconcile kube-apiserver network policy: %w", err)
-	}
-
-	// Reconcile openshift-monitoring Network Policy
-	policy = networkpolicy.OpenshiftMonitoringNetworkPolicy(controlPlaneNamespaceName)
-	if _, err := createOrUpdate(ctx, r.Client, policy, func() error {
-		return reconcileOpenshiftMonitoringNetworkPolicy(policy, hcluster)
-	}); err != nil {
-		return fmt.Errorf("failed to reconcile monitoring network policy: %w", err)
-	}
-
-	// Reconcile private-router Network Policy
-	if hcluster.Spec.Platform.Type == hyperv1.AWSPlatform {
-		policy = networkpolicy.PrivateRouterNetworkPolicy(controlPlaneNamespaceName)
-		if _, err := createOrUpdate(ctx, r.Client, policy, func() error {
-			return reconcilePrivateRouterNetworkPolicy(policy, hcluster)
-		}); err != nil {
-			return fmt.Errorf("failed to reconcile private router network policy: %w", err)
-		}
-	} else if hcluster.Spec.Platform.Type == hyperv1.KubevirtPlatform {
-		policy = networkpolicy.VirtLauncherNetworkPolicy(controlPlaneNamespaceName)
-		if _, err := createOrUpdate(ctx, r.Client, policy, func() error {
-			return reconcileVirtLauncherNetworkPolicy(policy, hcluster)
-		}); err != nil {
-			return fmt.Errorf("failed to reconcile virt launcher policy: %w", err)
-		}
-		egressFirewall := egressfirewall.VirtLauncherEgressFirewall(hcp.Namespace)
-		if _, err := createOrUpdate(ctx, r.Client, egressFirewall, func() error {
-			return reconcileVirtLauncherEgressFirewall(egressFirewall)
-		}); err != nil {
-			return fmt.Errorf("failed to reconcile firewall to deny metadata server egress: %w", err)
-		}
-	}
-
-	for _, svc := range hcluster.Spec.Services {
-		switch svc.Service {
-		case hyperv1.OAuthServer:
-			if svc.ServicePublishingStrategy.Type == hyperv1.NodePort {
-				// Reconcile nodeport-oauth Network Policy
-				policy = networkpolicy.NodePortOauthNetworkPolicy(controlPlaneNamespaceName)
-				if _, err := createOrUpdate(ctx, r.Client, policy, func() error {
-					return reconcileNodePortOauthNetworkPolicy(policy, hcluster)
-				}); err != nil {
-					return fmt.Errorf("failed to reconcile oauth server nodeport network policy: %w", err)
-				}
-			}
-		case hyperv1.Ignition:
-			if svc.ServicePublishingStrategy.Type == hyperv1.NodePort {
-				// Reconcile nodeport-ignition Network Policy
-				policy = networkpolicy.NodePortIgnitionNetworkPolicy(controlPlaneNamespaceName)
-				if _, err := createOrUpdate(ctx, r.Client, policy, func() error {
-					return reconcileNodePortIgnitionNetworkPolicy(policy, hcluster)
-				}); err != nil {
-					return fmt.Errorf("failed to reconcile ignition nodeport network policy: %w", err)
-				}
-			}
-		case hyperv1.Konnectivity:
-			if svc.ServicePublishingStrategy.Type == hyperv1.NodePort {
-				// Reconcile nodeport-konnectivity Network Policy
-				policy = networkpolicy.NodePortKonnectivityNetworkPolicy(controlPlaneNamespaceName)
-				if _, err := createOrUpdate(ctx, r.Client, policy, func() error {
-					return reconcileNodePortKonnectivityNetworkPolicy(policy, hcluster)
-				}); err != nil {
-					return fmt.Errorf("failed to reconcile konnectivity nodeport network policy: %w", err)
-				}
-			}
-		}
-	}
-
-	return nil
-}
-
 func (r *HostedClusterReconciler) validateConfigAndClusterCapabilities(ctx context.Context, hc *hyperv1.HostedCluster) error {
 	var errs []error
 	for _, svc := range hc.Spec.Services {
@@ -3624,301 +3519,6 @@ type ClusterMachineApproverConfig struct {
 }
 type NodeClientCert struct {
 	Disabled bool `json:"disabled,omitempty"`
-}
-
-func reconcileOpenshiftIngressNetworkPolicy(policy *networkingv1.NetworkPolicy) error {
-	policy.Spec.Ingress = []networkingv1.NetworkPolicyIngressRule{
-		{
-			From: []networkingv1.NetworkPolicyPeer{
-				{
-					NamespaceSelector: &metav1.LabelSelector{
-						MatchLabels: map[string]string{
-							"network.openshift.io/policy-group": "ingress",
-						},
-					},
-				},
-			},
-		},
-	}
-	policy.Spec.PodSelector = metav1.LabelSelector{}
-	policy.Spec.PolicyTypes = []networkingv1.PolicyType{networkingv1.PolicyTypeIngress}
-	return nil
-}
-
-func reconcileSameNamespaceNetworkPolicy(policy *networkingv1.NetworkPolicy) error {
-	policy.Spec.Ingress = []networkingv1.NetworkPolicyIngressRule{
-		{
-			From: []networkingv1.NetworkPolicyPeer{
-				{
-					PodSelector: &metav1.LabelSelector{},
-				},
-			},
-		},
-	}
-	policy.Spec.PodSelector = metav1.LabelSelector{}
-	policy.Spec.PolicyTypes = []networkingv1.PolicyType{networkingv1.PolicyTypeIngress}
-	return nil
-}
-
-func reconcileKASNetworkPolicy(policy *networkingv1.NetworkPolicy, hcluster *hyperv1.HostedCluster, isOpenShiftDNS bool, managementClusterNetwork *configv1.Network) error {
-	port := intstr.FromInt(int(hyperutil.BindAPIPortWithDefaultFromHostedCluster(hcluster, config.DefaultAPIServerPort)))
-	protocol := corev1.ProtocolTCP
-	policy.Spec.PolicyTypes = []networkingv1.PolicyType{networkingv1.PolicyTypeIngress}
-	policy.Spec.Ingress = []networkingv1.NetworkPolicyIngressRule{
-		{
-			From: []networkingv1.NetworkPolicyPeer{},
-			Ports: []networkingv1.NetworkPolicyPort{
-				{
-					Port:     &port,
-					Protocol: &protocol,
-				},
-			},
-		},
-	}
-	// We have to keep this in order to support 4.11 clusters where the KAS listen port == the external port
-	if hcluster.Spec.Networking.APIServer != nil && hcluster.Spec.Networking.APIServer.Port != nil {
-		externalPort := intstr.FromInt(int(*hcluster.Spec.Networking.APIServer.Port))
-		if port.IntValue() != externalPort.IntValue() {
-			policy.Spec.Ingress = append(policy.Spec.Ingress, networkingv1.NetworkPolicyIngressRule{
-				From: []networkingv1.NetworkPolicyPeer{},
-				Ports: []networkingv1.NetworkPolicyPort{
-					{
-						Port:     &externalPort,
-						Protocol: &protocol,
-					},
-				},
-			})
-		}
-	}
-
-	// NetworkPolicy egress is broken for 4.11 on OpenShiftSDN, this is a workaround
-	// https://issues.redhat.com/browse/OCPBUGS-2105
-	if managementClusterNetwork != nil && managementClusterNetwork.Spec.NetworkType != "OpenShiftSDN" {
-		// Allow traffic to same namespace
-		policy.Spec.Egress = []networkingv1.NetworkPolicyEgressRule{
-			{
-				To: []networkingv1.NetworkPolicyPeer{
-					{
-						PodSelector: &metav1.LabelSelector{},
-					},
-				},
-			},
-		}
-
-		if len(managementClusterNetwork.Spec.ClusterNetwork) > 0 {
-			clusterNetworks := []string{}
-			for _, network := range managementClusterNetwork.Spec.ClusterNetwork {
-				clusterNetworks = append(clusterNetworks, network.CIDR)
-			}
-
-			// Allow to any destination not on the management cluster service network
-			// i.e. block all inter-namespace egress from KAS not allowed by other rules
-			policy.Spec.Egress = append(policy.Spec.Egress,
-				networkingv1.NetworkPolicyEgressRule{
-					To: []networkingv1.NetworkPolicyPeer{
-						{
-							IPBlock: &networkingv1.IPBlock{
-								CIDR:   "0.0.0.0/0",
-								Except: clusterNetworks,
-							},
-						},
-					},
-				})
-		}
-
-		if isOpenShiftDNS {
-			// Allow traffic to openshift-dns namespace
-			dnsUDPPort := intstr.FromInt(5353)
-			dnsUDPProtocol := corev1.ProtocolUDP
-			dnsTCPPort := intstr.FromInt(5353)
-			dnsTCPProtocol := corev1.ProtocolTCP
-			policy.Spec.Egress = append(policy.Spec.Egress, networkingv1.NetworkPolicyEgressRule{
-				To: []networkingv1.NetworkPolicyPeer{
-					{
-						NamespaceSelector: &metav1.LabelSelector{
-							MatchLabels: map[string]string{
-								"kubernetes.io/metadata.name": "openshift-dns",
-							},
-						},
-					},
-				},
-				Ports: []networkingv1.NetworkPolicyPort{
-					{
-						Port:     &dnsUDPPort,
-						Protocol: &dnsUDPProtocol,
-					},
-					{
-						Port:     &dnsTCPPort,
-						Protocol: &dnsTCPProtocol,
-					},
-				},
-			})
-		} else {
-			// All traffic to any destination on port 53 for both TCP and UDP
-			dnsUDPPort := intstr.FromInt(53)
-			dnsUDPProtocol := corev1.ProtocolUDP
-			dnsTCPPort := intstr.FromInt(53)
-			dnsTCPProtocol := corev1.ProtocolTCP
-			policy.Spec.Egress = append(policy.Spec.Egress, networkingv1.NetworkPolicyEgressRule{
-				Ports: []networkingv1.NetworkPolicyPort{
-					{
-						Port:     &dnsUDPPort,
-						Protocol: &dnsUDPProtocol,
-					},
-					{
-						Port:     &dnsTCPPort,
-						Protocol: &dnsTCPProtocol,
-					},
-				},
-			})
-		}
-		policy.Spec.PolicyTypes = append(policy.Spec.PolicyTypes, networkingv1.PolicyTypeEgress)
-	}
-
-	policy.Spec.PodSelector = metav1.LabelSelector{
-		MatchLabels: map[string]string{
-			"app": "kube-apiserver",
-		},
-	}
-
-	return nil
-}
-
-func reconcilePrivateRouterNetworkPolicy(policy *networkingv1.NetworkPolicy, hcluster *hyperv1.HostedCluster) error {
-	httpPort := intstr.FromInt(8080)
-	httpsPort := intstr.FromInt(8443)
-	protocol := corev1.ProtocolTCP
-	policy.Spec.Ingress = []networkingv1.NetworkPolicyIngressRule{
-		{
-			From: []networkingv1.NetworkPolicyPeer{},
-			Ports: []networkingv1.NetworkPolicyPort{
-				{
-					Port:     &httpPort,
-					Protocol: &protocol,
-				},
-				{
-					Port:     &httpsPort,
-					Protocol: &protocol,
-				},
-			},
-		},
-	}
-	policy.Spec.PodSelector = metav1.LabelSelector{
-		MatchLabels: map[string]string{
-			"app": "private-router",
-		},
-	}
-	policy.Spec.PolicyTypes = []networkingv1.PolicyType{networkingv1.PolicyTypeIngress}
-	return nil
-}
-
-func reconcileNodePortOauthNetworkPolicy(policy *networkingv1.NetworkPolicy, hcluster *hyperv1.HostedCluster) error {
-	port := intstr.FromInt(6443)
-	protocol := corev1.ProtocolTCP
-	policy.Spec.Ingress = []networkingv1.NetworkPolicyIngressRule{
-		{
-			From: []networkingv1.NetworkPolicyPeer{},
-			Ports: []networkingv1.NetworkPolicyPort{
-				{
-					Port:     &port,
-					Protocol: &protocol,
-				},
-			},
-		},
-	}
-	policy.Spec.PodSelector = metav1.LabelSelector{
-		MatchLabels: map[string]string{
-			"app": "oauth-openshift",
-		},
-	}
-	policy.Spec.PolicyTypes = []networkingv1.PolicyType{networkingv1.PolicyTypeIngress}
-	return nil
-}
-
-func reconcileNodePortIgnitionNetworkPolicy(policy *networkingv1.NetworkPolicy, hcluster *hyperv1.HostedCluster) error {
-	port := intstr.FromInt(9090)
-	protocol := corev1.ProtocolTCP
-	policy.Spec.Ingress = []networkingv1.NetworkPolicyIngressRule{
-		{
-			From: []networkingv1.NetworkPolicyPeer{},
-			Ports: []networkingv1.NetworkPolicyPort{
-				{
-					Port:     &port,
-					Protocol: &protocol,
-				},
-			},
-		},
-	}
-	policy.Spec.PodSelector = metav1.LabelSelector{
-		MatchLabels: map[string]string{
-			"app": "ignition-server",
-		},
-	}
-	policy.Spec.PolicyTypes = []networkingv1.PolicyType{networkingv1.PolicyTypeIngress}
-	return nil
-}
-
-func reconcileNodePortKonnectivityNetworkPolicy(policy *networkingv1.NetworkPolicy, hcluster *hyperv1.HostedCluster) error {
-	port := intstr.FromInt(8091)
-	protocol := corev1.ProtocolTCP
-	policy.Spec.Ingress = []networkingv1.NetworkPolicyIngressRule{
-		{
-			From: []networkingv1.NetworkPolicyPeer{},
-			Ports: []networkingv1.NetworkPolicyPort{
-				{
-					Port:     &port,
-					Protocol: &protocol,
-				},
-			},
-		},
-	}
-	policy.Spec.PodSelector = metav1.LabelSelector{
-		MatchLabels: map[string]string{
-			"app": "konnectivity-server",
-		},
-	}
-	policy.Spec.PolicyTypes = []networkingv1.PolicyType{networkingv1.PolicyTypeIngress}
-	return nil
-}
-
-func reconcileOpenshiftMonitoringNetworkPolicy(policy *networkingv1.NetworkPolicy, hcluster *hyperv1.HostedCluster) error {
-	policy.Spec.Ingress = []networkingv1.NetworkPolicyIngressRule{
-		{
-			From: []networkingv1.NetworkPolicyPeer{
-				{
-					NamespaceSelector: &metav1.LabelSelector{
-						MatchLabels: map[string]string{
-							"network.openshift.io/policy-group": "monitoring",
-						},
-					},
-				},
-			},
-		},
-	}
-	policy.Spec.PodSelector = metav1.LabelSelector{}
-	policy.Spec.PolicyTypes = []networkingv1.PolicyType{networkingv1.PolicyTypeIngress}
-	return nil
-}
-
-func reconcileVirtLauncherNetworkPolicy(policy *networkingv1.NetworkPolicy, hcluster *hyperv1.HostedCluster) error {
-	protocolTCP := corev1.ProtocolTCP
-	protocolUDP := corev1.ProtocolUDP
-	protocolSCTP := corev1.ProtocolSCTP
-	policy.Spec.Ingress = []networkingv1.NetworkPolicyIngressRule{
-		{
-			Ports: []networkingv1.NetworkPolicyPort{
-				{Protocol: &protocolTCP},
-				{Protocol: &protocolUDP},
-				{Protocol: &protocolSCTP},
-			},
-		},
-	}
-	policy.Spec.PodSelector = metav1.LabelSelector{
-		MatchLabels: map[string]string{
-			"kubevirt.io": "virt-launcher",
-		},
-	}
-	return nil
 }
 
 func reconcileVirtLauncherEgressFirewall(egressFirewall *unstructured.Unstructured) error {
